requirements/dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/dev.in -o requirements/dev.txt
aiohttp==3.9.3
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.8.3
    # via twilio
aiosignal==1.3.1
    # via aiohttp
amqp==5.2.0
    # via kombu
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
asttokens==2.4.1
    # via stack-data
async-timeout==4.0.3
    # via redis
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
billiard==4.2.0
    # via celery
black==24.2.0
celery==5.4.0rc1
    # via flower
certifi==2024.2.2
    # via
    #   django-healthchecks
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
decorator==5.1.1
    # via ipython
distlib==0.3.8
    # via virtualenv
dj-database-url==2.1.0
django==5.0.2
    # via
    #   dj-database-url
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-healthchecks
    #   django-linear-migrations
    #   django-localflavor
    #   django-phonenumber-field
    #   django-redis
    #   django-rich
    #   django-version-checks
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   drf-standardized-errors
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-healthchecks==1.5.0
django-linear-migrations==2.12.0
django-localflavor==4.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-rich==1.8.0
django-split-settings==1.2.0
django-version-checks==1.2.0
djangorestframework==3.14.0
    # via
    #   djangorestframework-simplejwt
    #   drf-extensions
    #   drf-spectacular
    #   drf-standardized-errors
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
drf-spectacular==0.27.1
drf-standardized-errors==0.13.0
executing==2.0.1
    # via stack-data
factory-boy==3.3.0
faker==23.3.0
    # via factory-boy
filelock==3.13.1
    # via virtualenv
flake8==7.0.0
    # via flake8-pyproject
flake8-pyproject==1.2.3
flower==2.0.1
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gunicorn==21.2.0
humanize==4.9.0
    # via flower
identify==2.5.35
    # via pre-commit
idna==3.6
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
ipython==8.22.1
jedi==0.19.1
    # via ipython
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.5
    # via celery
markdown-it-py==3.0.0
    # via rich
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
packaging==23.2
    # via
    #   black
    #   django-version-checks
    #   gunicorn
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pillow==10.2.0
pip-lock==2.11.0
platformdirs==4.2.0
    # via
    #   black
    #   virtualenv
pluggy==1.4.0
    # via pytest
pre-commit==3.6.2
prometheus-client==0.20.0
    # via flower
prompt-toolkit==3.0.43
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.9
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pygments==2.17.2
    # via
    #   ipython
    #   rich
pyjwt==2.8.0
    # via
    #   djangorestframework-simplejwt
    #   twilio
pyotp==2.9.0
pytest==8.0.2
    # via pytest-django
pytest-django==4.8.0
python-dateutil==2.8.2
    # via
    #   celery
    #   faker
    #   time-machine
python-dotenv==1.0.1
python-stdnum==1.19
    # via django-localflavor
pytz==2024.1
    # via
    #   djangorestframework
    #   flower
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   pre-commit
redis==5.0.2
    # via django-redis
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   django-healthchecks
    #   twilio
rich==13.7.0
    # via django-rich
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
ruff==0.3.1
setuptools==69.1.1
    # via nodeenv
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
time-machine==2.13.0
tornado==6.4
    # via flower
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
twilio==8.13.0
typing-extensions==4.10.0
    # via dj-database-url
tzdata==2024.1
    # via celery
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.1
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.25.1
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.9.4
    # via aiohttp
